0.12 compatibility/namespace refactor ep. 1

+++ src/pocketmine/network/protocol/TileEntityDataPacket.php
@@ -24,14 +24,6 @@ namespace pocketmine\network\protocol;
 use pocketmine\utils\Binary;
 
 
-
-
-
-
-
-
-
-
 class TileEntityDataPacket extends DataPacket{
 	public static $pool = [];
 	public static $next = 0;
@@ -47,7 +39,7 @@ class TileEntityDataPacket extends DataPacket{
 
 	public function decode(){
 		$this->x = $this->getInt();
-		$this->y = $this->getByte();
+		$this->y = $this->getInt();
 		$this->z = $this->getInt();
 		$this->namedtag = $this->get(true);
 	}
@@ -55,7 +47,7 @@ class TileEntityDataPacket extends DataPacket{
 	public function encode(){
 		$this->reset();
 		$this->putInt($this->x);
-		$this->putByte($this->y);
+		$this->putInt($this->y);
 		$this->putInt($this->z);
 		$this->put($this->namedtag);
 	}
